hardware/orangepi/pyproject.toml
[project]
name = "hardware"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "adafruit-circuitpython-bno055>=5.4.21",
    "flask>=3.1.3",
    "gpiod>=2.4.2",
    "opi-gpio>=0.5.2",
    "pynmea2>=1.19.0",
    "pyserial>=3.5",
]
